Refactor of the booking method - all tested
