+++ tee-time-core/tee-time-core-api/src/main/java/com/golf/teetimecoreapi/service/BookingService.java
@@ -1,106 +1,180 @@
-//package com.golf.teetimecoreapi.service;
-//
-//import com.golf.teetimecoreapi.model.BookingRequest;
-//import com.golf.teetimecoreapi.model.BookingResponse;
-//import com.golf.teetimecoreapi.exception.BookingException;
-//import lombok.extern.slf4j.Slf4j;
-//import org.openqa.selenium.By;
-//import org.openqa.selenium.TimeoutException;
-//import org.openqa.selenium.WebElement;
-//import org.openqa.selenium.support.ui.ExpectedConditions;
-//import org.openqa.selenium.support.ui.Select;
-//import org.openqa.selenium.support.ui.WebDriverWait;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.beans.factory.annotation.Value;
-//import org.springframework.stereotype.Service;
-//import org.openqa.selenium.WebDriver;
-//
-//import java.time.Duration;
-//
-//@Service
-//@Slf4j
-//public class BookingService {
-//
-//    private final WebDriver webDriver;
-//
-//    @Value("${golf.booking.base-url}")
-//    private String baseUrl;
-//
-//    @Autowired
-//    public BookingService(WebDriver webDriver) {
-//        this.webDriver = webDriver;
-//    }
-//
-//    public BookingResponse makeBooking(BookingRequest request) {
-//        try {
-//            String courseUrl = baseUrl + "/course/" + request.getCourse();
-//            webDriver.get(courseUrl);
-//
-//            // Execute booking logic
-//            boolean success = executeBookingFlow(request);
-//
-//            if (success) {
-//                String confirmationNumber = extractConfirmationNumber();
-//                return BookingResponse.builder()
-//                        .status("CONFIRMED")
-//                        .confirmationNumber(confirmationNumber)
-//                        .build();
-//            } else {
-//                return BookingResponse.builder()
-//                        .status("FAILED")
-//                        .message("Unable to complete booking")
-//                        .build();
-//            }
-//        } catch (Exception e) {
-//            log.error("Booking failed", e);
-//            throw new BookingException("Failed to make booking: " + e.getMessage());
-//        }
-//    }
-//
-//    private boolean executeBookingFlow(BookingRequest request) {
-//        try {
-//            // Select date
-//            WebElement dateInput = findElement(By.id("date"));
-//            dateInput.sendKeys(request.getDate());
-//
-//            // Select time
-//            WebElement timeSlot = findElement(By.xpath("//div[@data-time='" + request.getTime() + "']"));
-//            timeSlot.click();
-//
-//            // Set players
-//            WebElement playerSelect = findElement(By.id("players"));
-//            Select select = new Select(playerSelect);
-//            select.selectByValue(String.valueOf(request.getPlayers()));
-//
-//            // Confirm booking
-//            WebElement confirmButton = findElement(By.id("confirm-booking"));
-//            confirmButton.click();
-//
-//            // Wait for confirmation
-//            return waitForConfirmation();
-//        } catch (Exception e) {
-//            log.error("Error during booking flow", e);
-//            return false;
-//        }
-//    }
-//
-//    private WebElement findElement(By by) {
-//        return new WebDriverWait(webDriver, Duration.ofSeconds(10))
-//                .until(ExpectedConditions.elementToBeClickable(by));
-//    }
-//
-//    private boolean waitForConfirmation() {
-//        try {
-//            WebElement confirmation = new WebDriverWait(webDriver, Duration.ofSeconds(15))
-//                    .until(ExpectedConditions.presenceOfElementLocated(By.className("booking-confirmation")));
-//            return confirmation.isDisplayed();
-//        } catch (TimeoutException e) {
-//            return false;
-//        }
-//    }
-//
-//    private String extractConfirmationNumber() {
-//        WebElement confirmationElement = findElement(By.className("confirmation-number"));
-//        return confirmationElement.getText();
-//    }
-//}
+package com.golf.teetimecoreapi.service;
+
+import lombok.extern.slf4j.Slf4j;
+import org.openqa.selenium.By;
+import org.openqa.selenium.TimeoutException;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.springframework.stereotype.Service;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+
+import java.text.SimpleDateFormat;
+import java.util.Arrays;
+import java.util.Date;
+import java.util.List;
+
+@Service
+@Slf4j
+public class BookingService {
+
+    public static final Log LOGGER = LogFactory.getLog(BookingService.class);
+
+    public WebElement findFinaliseButton(WebDriver driver, WebDriverWait wait) {
+        List<By> locators = Arrays.asList(
+                By.id("MainContent_FinishButton"),
+                By.xpath("//input[@value='FINALISE BOOKING']"),
+                By.cssSelector(".module-button-cta-nzg.btn.btn-primary"),
+                By.name("ctl00$MainContent$FinishButton"),
+                By.xpath("//div[contains(@class, 'alert-info')]//input[@type='submit']")
+        );
+
+        for (By locator : locators) {
+            try {
+                return wait.until(ExpectedConditions.elementToBeClickable(locator));
+            } catch (TimeoutException e) {
+                LOGGER.error("Could not find finalise button with locator: " + locator);
+                continue;
+            }
+        }
+        throw new RuntimeException("Could not find finalise booking button");
+    }
+
+    public WebElement findSubmitButton(WebDriver driver, WebDriverWait wait) {
+        List<By> locators = Arrays.asList(
+                By.id("MainContent_ContinueButton"),
+                By.xpath("//input[@type='submit' and @value='SUBMIT']"),
+                By.cssSelector("input.module-button-cta.btn.btn-primary"),
+                By.xpath("//input[contains(@class, 'module-button-cta') and contains(@class, 'btn-primary')]"),
+                By.name("ctl00$MainContent$ContinueButton")
+        );
+
+        for (By locator : locators) {
+            try {
+                return wait.until(ExpectedConditions.elementToBeClickable(locator));
+            } catch (TimeoutException e) {
+                LOGGER.error("Could not find submit button with locator: " + locator);
+                continue;
+            }
+        }
+        throw new RuntimeException("Could not find submit button");
+    }
+
+    public WebElement findTimeSlot(WebDriver driver, WebDriverWait wait, String time) {
+        List<By> locators = Arrays.asList(
+                // Find by exact time and "Book Here" link
+                By.xpath(String.format("//td[contains(@class, 'xtime') and text()='%s']/..//a[contains(@class, 'book_here_link')]", time)),
+                // Alternative approach using ancestor
+                By.xpath(String.format("//td[@class='xtime'][text()='%s']/following-sibling::td//a[contains(@class, 'book_here_link')]", time))
+        );
+
+        for (By locator : locators) {
+            try {
+                return wait.until(ExpectedConditions.elementToBeClickable(locator));
+            } catch (TimeoutException e) {
+                LOGGER.error("Could not find time slot with locator: " + locator);
+                continue;
+            }
+        }
+        throw new RuntimeException("Could not find booking link for time: " + time);
+    }
+
+    public WebElement findDateCell(WebDriver driver, WebDriverWait wait, WebElement courseRow, Date date) {
+        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+        String formattedDate = dateFormat.format(date);
+
+        try {
+            // Look for available date cells only within the course row
+            return courseRow.findElement(
+                    By.xpath(String.format(".//td[contains(@class, 'available') and contains(@onmouseover, '%s')]", formattedDate))
+            );
+        } catch (Exception e) {
+            LOGGER.error("Could not find date cell for " + formattedDate + " in course row");
+            throw new RuntimeException("Could not find available date cell for: " + formattedDate);
+        }
+    }
+
+    public WebElement findCourseElement(WebDriver driver, String course) {
+        List<WebElement> courseElements = driver.findElements(By.cssSelector(".club_info.course-name span"));
+        for (WebElement courseElement : courseElements) {
+            if (courseElement.getText().equalsIgnoreCase(course)) {
+                return courseElement;
+            }
+        }
+        return null;
+    }
+
+    public WebElement findMembershipNumberField(WebDriver driver, WebDriverWait wait) {
+        List<By> locators = Arrays.asList(
+                By.xpath("//input[contains(@class, 'form-control') and contains(@id, 'tbMembershipNumber')]"),
+                By.cssSelector("input[placeholder='e.g. 12343243']"),
+                By.name("ctl*$tbMembershipNumber")  // Partial name matching
+        );
+
+        for (By locator : locators) {
+            try {
+                return wait.until(ExpectedConditions.visibilityOfElementLocated(locator));
+            } catch (TimeoutException e) {
+                // Log error finding element and continue to next locator
+                LOGGER.error("Could not find membership number input field with locator: " + locator);
+                return null;
+            }
+        }
+        throw new RuntimeException("Could not find membership number input field");
+    }
+
+    public WebElement findPasswordField(WebDriver driver, WebDriverWait wait) {
+        List<By> locators = Arrays.asList(
+                By.xpath("//input[contains(@class, 'form-control') and contains(@type, 'password')]"),
+                By.cssSelector("input[type='password']"),
+                By.name("ctl*$tbPassword"),
+                By.xpath("//label[text()='Password']/following-sibling::input[@type='password']")
+        );
+
+        for (By locator : locators) {
+            try {
+                return wait.until(ExpectedConditions.visibilityOfElementLocated(locator));
+            } catch (TimeoutException e) {
+                LOGGER.error("Could not find password input field with locator: " + locator);
+                return null;
+            }
+        }
+        throw new RuntimeException("Could not find password input field");
+    }
+
+    public WebElement findLoginButton(WebDriver driver, WebDriverWait wait) {
+        List<By> locators = Arrays.asList(
+                By.xpath("//input[@type='submit' and @value='Login']"),
+                By.cssSelector("input.btn-primary.btn-block[value='Login']"),
+                By.name("ctl*$btnLogin"),
+                By.id("ctl61_btnLogin"),
+                By.xpath("//input[contains(@class, 'btn-primary') and contains(@class, 'btn-block')]")
+        );
+
+        for (By locator : locators) {
+            try {
+                return wait.until(ExpectedConditions.elementToBeClickable(locator));
+            } catch (TimeoutException e) {
+                LOGGER.error("Could not find login button with locator: " + locator);
+                return null;
+            }
+        }
+        throw new RuntimeException("Could not find login button");
+    }
+
+    public void nzGolfLogin(WebDriver driver, WebDriverWait wait) {
+        // Login procedure
+        WebElement membershipNumberField = findMembershipNumberField(driver, wait);
+        WebElement passwordField = findPasswordField(driver, wait);
+        WebElement loginButton = findLoginButton(driver, wait);
+
+        // Enter login details
+
+        membershipNumberField.sendKeys("4676144"); // NewUserConfiguration.getMembershipNumber());
+        passwordField.sendKeys("Sard3nia");  //NewUserConfiguration.getGolfPassword()); // *******
+        loginButton.click();
+    }
+
+}
